delete src/ui/control/attribution_control.test.ts add test/unit/ui/control/
